requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
arrow==1.2.3
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
slack-sdk==3.19.5
todoist-api-python==2.0.2
urllib3==1.26.13
yarl==1.8.2
